% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComByPostal.R
\name{ComByPostal}
\alias{ComByPostal}
\title{Get City by Postal Code}
\usage{
ComByPostal(codePostal)
}
\arguments{
\item{codePostal}{a numeric vector with a postal code.}
}
\value{
Returns a data.frame with name, INSEE code, postal code, INSEE department code, INSEE region code, population (approx), surface (in hectares), lat and long (WGS-84).
}
\description{
Get informations about a French city by its INSEE Code. Please note that this package works only with French cities.
}
\details{
Takes a postal code, returns a data.frame with the available values.
}
\note{
If you don't know the postal code of the city you're looking for, you can find it by using the \code{\link{ComByName}} function.
}
\examples{
ComByPostal(codePostal = 29000)
}

