% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepByName.R
\name{DepByName}
\alias{DepByName}
\title{Get Department by Name}
\usage{
DepByName(nom)
}
\arguments{
\item{nom}{a character string with the name of the department.}
}
\value{
Returns a data.frame with name, INSEE code, and region code of the department.
}
\description{
Get informations about a French Department by its name. Please note that this package works only with French Department.
}
\details{
Takes a department INSEE code, returns a data.frame with the available values. Partial matches are possible. In that case, typographic pertinence scores are given.
}
\examples{
DepByName(nom = "Ille-et-Vilaine")
DepByName(nom = "mo")
}

