% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepByReg.R
\name{DepByReg}
\alias{DepByReg}
\title{Get Departments by Region}
\usage{
DepByReg(codeRegion)
}
\arguments{
\item{codeRegion}{a numeric vector with an INSEE code.}
}
\value{
Returns a data.frame with departments names, INSEE codes, and region codes of the departments.
}
\description{
Get informations about all the departments in a French region by its INSEE code. Please note that this package works only with French departments.
}
\details{
Takes a region INSEE code, returns a data.frame with the available values.
}
\note{
If you don't know the INSEE code of the region you're looking for, you can find it by using the \code{\link{RegByName}} function.
}
\examples{
DepByReg(codeRegion = 53)
}

