% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeinsee.R
\name{getByIC}
\alias{getByIC}
\title{Get By INSEE Code}
\usage{
getByIC(codeInsee)
}
\arguments{
\item{an}{INSEE Code}
}
\value{
a data.frame with names, postal code, INSEE code, surface, lat and long.
}
\description{
Takes an INSEE Code, return a data.frame with name, postal code, INSEE code, surface, lat and long of the city.
}
\examples{
getByIC(codeInsee = "88160") 
getByIC(codeInsee = "31555")
library(plyr)
ldply(c("88160", "60057"), getByIC)
}

