% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ville.R
\name{getByName}
\alias{getByName}
\title{Get by name}
\usage{
getByName(Name)
}
\arguments{
\item{Name}{the name of a commune}
}
\value{
a data.frame with names, postal code, INSEE code, surface, lat and long.
}
\description{
Takes a the name of a French commune and returns name, code Insee, postal code, area, latitude and longitude
}
\examples{
getByName("Épinal")

}

