% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codepostal.R
\name{getByPC}
\alias{getByPC}
\title{Get By Postal Code}
\usage{
getByPC(codePostal)
}
\arguments{
\item{an}{postal code}
}
\value{
a data.frame with names, postal code, INSEE code, surface, lat and long.
}
\description{
Takes a french postal code, return a data.frame with name, postal code, INSEE code, surface, lat and long of the city.
}
\examples{
getByPC("88000")
}

