% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegByCode.R
\name{RegByCode}
\alias{RegByCode}
\title{Get Region by INSEE Code}
\usage{
RegByCode(codeRegion)
}
\arguments{
\item{codeRegion}{a numeric vector with an INSEE code.}
}
\value{
Returns a data.frame with name and region code of the region.
}
\description{
Get informations about a french region by its INSEE code. Please note that this package works only with french regions.
}
\details{
Takes an INSEE code, returns a data.frame with the available values.
}
\examples{
RegByCode(53)
}

