% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegByName.R
\name{RegByName}
\alias{RegByName}
\title{Get Region by Name}
\usage{
RegByName(nom)
}
\arguments{
\item{nom}{a character string with the name of the department. Partial matches are possible. In that case, typographic pertinence scores are given.}
}
\value{
Returns a data.frame with name, INSEE code, and typographic pertinence score for the request.
}
\description{
Get informations about a French region by its name. Please note that this package works only with French regions.
}
\details{
Takes a name, returns a data.frame with the available values.
}
\examples{
RegByName(nom = "Bretagne")
RegByName(nom = "M")
}

