% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComByName.R
\name{ComByName}
\alias{ComByName}
\title{Get City by Name}
\usage{
ComByName(nom, boost = FALSE)
}
\arguments{
\item{nom}{a character string with the name of the city.}

\item{boost}{a TRUE or FALSE. Default is FALSE. If TRUE, typographic pertinence score will be weighted by population.}
}
\value{
Returns a data.frame with name(s), INSEE code(s), postal code(s), INSEE department code(s), INSEE region code(s), population (approx), surface(s) (in hectares), lat and long (WGS-84).
}
\description{
Get informations about a french city by its name (partial matches possible). Please note that this package works only with french cities.
}
\details{
Takes a the name of a French city, returns a data.frame with the available values. Partial matches are possible. In that case, typographic pertinence scores are given, and can be weighted by population with the "boost" argument.
}
\examples{
ComByName("Brest")
ComByName("Vitré", boost = TRUE)
ComByName("Lo")
}

