% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ville.R
\name{getByName}
\alias{getByName}
\title{Get By Name}
\usage{
getByName(Name)
}
\arguments{
\item{a}{name, as characters}
}
\value{
a data.frame with names, postal code, INSEE code, surface, lat and long.
}
\description{
Takes the name of a french city, return a data.frame with name, postal code, INSEE code, surface, lat and long of the city.
}

